+++ openfisca_survey_manager/scenarios.py
@@ -5,7 +5,8 @@ from __future__ import division
 import humanize
 import logging
 import numpy as np
-import pandas
+import pandas as pd
+import re
 
 
 from openfisca_core import formulas, periods, simulations
@@ -113,47 +114,41 @@ class AbstractSurveyScenario(object):
         elif aggfunc == 'count':
             return (weight * filter_dummy).sum()
 
-    def compute_pivot_table(
-            self, aggfunc = 'mean', columns = None, difference = None, filter_by = None, index = None,
+    def compute_pivot_table(self, aggfunc = 'mean', columns = None, difference = False, filter_by = None, index = None,
             period = None, reference = False, values = None, missing_variable_default_value = np.nan):
         assert aggfunc in ['count', 'mean', 'sum']
+        assert columns or index or values
+        assert not (difference and reference), "Can't have difference and reference both set to True"
 
         tax_benefit_system = self.tax_benefit_system
+
+        if isinstance(columns, str):
+            columns = [columns]
+        elif columns is None:
+            columns = []
+        assert isinstance(columns, list)
+
+        if isinstance(index, str):
+            index = [index]
+        elif index is None:
+            index = []
+        assert isinstance(index, list)
+
+        if isinstance(values, str):
+            values = [values]
+        elif values is None:
+            values = []
+        assert isinstance(values, list)
+
+        entity_key = None
+        for axe in [columns, index, values]:
+            if axe is not None and entity_key is None:
+                entity_key = tax_benefit_system.column_by_name[axe[0]].entity.key
+
         if filter_by is None and self.filtering_variable_by_entity is not None:
-            entity_key = tax_benefit_system.column_by_name[values[0]].entity.key
             filter_by = self.filtering_variable_by_entity.get(entity_key)
 
-        assert isinstance(values, (str, list))
-        if isinstance(values, str):
-            values = ['values']
-
-        # assert len(values) == 1
-
-        if difference:
-            return (
-                self.compute_pivot_table(
-                    aggfunc = aggfunc, columns = columns, filter_by = filter_by, index = index,
-                    period = period, reference = False, values = values,
-                    missing_variable_default_value = missing_variable_default_value
-                    ) -
-                self.compute_pivot_table(
-                    aggfunc = aggfunc, columns = columns, filter_by = filter_by, index = index,
-                    period = period, reference = True, values = values,
-                    missing_variable_default_value = missing_variable_default_value)
-                )
-
-        if reference:
-            simulation = self.reference_simulation
-        else:
-            simulation = self.simulation
-
-        assert simulation is not None
-
-        index_list = index if index is not None else []
-        columns_list = columns if columns is not None else []
-        variables = set(index_list + values + columns_list)
-        entity_key = tax_benefit_system.column_by_name[values[0]].entity.key
-
+        variables = set(index + values + columns)
         # Select the entity weight corresponding to the variables that will provide values
         if self.weight_column_name_by_entity is not None:
             weight = self.weight_column_name_by_entity[entity_key]
@@ -173,66 +168,172 @@ class AbstractSurveyScenario(object):
                     entity_key,
                     )
 
-        def calculate_variable(var):
-            if var in simulation.tax_benefit_system.column_by_name:
-                return simulation.calculate_add(var, period = period)
-            else:
-                log.info("Variable {} not found. Assiging {}".format(variable, missing_variable_default_value))
-                return missing_variable_default_value
-
-        data_frame = pandas.DataFrame(dict(
-            (variable, calculate_variable(variable)) for variable in variables
-            ))
+        if difference:
+            data_frame = (
+                self.create_data_frame_by_entity(
+                    values, period = period, reference = False, index = False)[entity_key] -
+                self.create_data_frame_by_entity(
+                    values, period = period, reference = True, index = False)[entity_key]
+                )
+        else:
+            data_frame = (
+                self.create_data_frame_by_entity(
+                    values, period = period, reference = reference, index = False)[entity_key]
+                )
+        additionnal_variables = []
+        if filter_by is not None:
+            additionnal_variables.append(filter_by)
+        if weight is not None:
+            additionnal_variables.append(weight)
+        reference_variables = set(index + columns + additionnal_variables)
+        data_frame = pd.concat(
+            [
+                self.create_data_frame_by_entity(
+                    variables = reference_variables, period = period, reference = True, index = False
+                    )[entity_key],
+                data_frame,
+                ],
+            axis = 1,
+            )
         if filter_by in data_frame:
-            filter_dummy = data_frame.get(filter_by)
+            filter_dummy = data_frame[filter_by]
+
         if weight is None:
             weight = 'weight'
             data_frame[weight] = 1.0
 
-        data_frame[values[0]] = data_frame[values[0]] * data_frame[weight] * filter_dummy
-        pivot_sum = data_frame.pivot_table(index = index, columns = columns, values = values, aggfunc = 'sum')
-        pivot_mass = data_frame.pivot_table(index = index, columns = columns, values = weight, aggfunc = 'sum')
-        if aggfunc == 'mean':
-            return (pivot_sum / pivot_mass)
-        elif aggfunc == 'sum':
-            return pivot_sum
-        elif aggfunc == 'count':
-            return pivot_mass
+        data_frame[weight] = data_frame[weight] * filter_dummy
 
-    def create_data_frame_by_entity(self, variables = None, indices = False, period = None, reference = False,
-            roles = False):
-        assert variables is not None or indices or roles
-        tax_benefit_system = self.tax_benefit_system
+        if values:
+            data_frame_by_value = dict()
+            for value in values:
+                data_frame[value] = data_frame[value] * data_frame[weight]
+                data_frame[value].fillna(missing_variable_default_value)
+                pivot_sum = data_frame.pivot_table(index = index, columns = columns, values = values, aggfunc = 'sum')
+                pivot_mass = data_frame.pivot_table(index = index, columns = columns, values = weight, aggfunc = 'sum')
+                if aggfunc == 'mean':
+                    result = (pivot_sum / pivot_mass.loc[weight])
+                elif aggfunc == 'sum':
+                    result = pivot_sum
+                elif aggfunc == 'count':
+                    result = pivot_mass
 
-        if reference:
-            simulation = self.reference_simulation
+                data_frame_by_value[value] = result
+
+            if len(data_frame_by_value.keys()) > 1:
+                return data_frame_by_value
+            else:
+                return data_frame_by_value.values()[0]
+
         else:
-            simulation = self.simulation
+            assert aggfunc == 'count', "Can only use count for aggfunc if no values"
+            return data_frame.pivot_table(index = index, columns = columns, values = weight, aggfunc = 'sum')
 
-        assert simulation is not None
+    def create_data_frame_by_entity(self, variables = None, expressions = None, filter_by = None, index = False,
+            period = None, reference = False, merge = False, ignore_missing_variables = False):
 
-        missing_variables = set(variables).difference(set(self.tax_benefit_system.column_by_name.keys()))
+        simulation = self.reference_simulation if (reference and self.reference_simulation) else self.simulation
+        tax_benefit_system = self.reference_tax_benefit_system if (reference and self.reference_tax_benefit_system) else self.tax_benefit_system
+
+        assert variables or index or expressions or filter_by
+
+        if merge:
+            index = True
+        if expressions is None:
+            expressions = []
+
+        if filter_by is not None:
+            if filter_by in tax_benefit_system.column_by_name.keys():
+                variables.append(filter_by)
+                filter_entity_key = tax_benefit_system.column_by_name.get(filter_by).entity.key
+            else:
+                filter_entity_key = assert_variables_in_same_entity(self, get_words(filter_by))
+                expressions.append(filter_by)
+
+        expressions_by_entity_key = dict()
+        for expression in expressions:
+            expression_variables = get_words(expression)
+            entity_key = assert_variables_in_same_entity(self, expression_variables)
+            if entity_key in expressions_by_entity_key:
+                expressions_by_entity_key[entity_key].append(expression)
+            else:
+                expressions_by_entity_key[entity_key] = [expression]
+            variables += expression_variables
+
+        variables = set(variables)
+
+        missing_variables = set(variables).difference(set(tax_benefit_system.column_by_name.keys()))
         if missing_variables:
             log.info("These variables aren't par of the tax-benefit system: {}".format(missing_variables))
         columns_to_fetch = [
-            self.tax_benefit_system.column_by_name.get(variable_name) for variable_name in variables
-            if self.tax_benefit_system.column_by_name.get(variable_name) is not None
+            tax_benefit_system.column_by_name.get(variable_name) for variable_name in variables
+            if tax_benefit_system.column_by_name.get(variable_name) is not None
             ]
+
+        assert len(columns_to_fetch) >= 1, "None of the requested variables {} are in the tax-benefit-system".format(
+            variables)
+
+        assert simulation is not None
+
         openfisca_data_frame_by_entity_key = dict()
+        non_person_entities = list()
+
         for entity in tax_benefit_system.entities:
             entity_key = entity.key
             column_names = [
                 column.name for column in columns_to_fetch
                 if column.entity == entity
                 ]
-            openfisca_data_frame_by_entity_key[entity_key] = pandas.DataFrame(
+            openfisca_data_frame_by_entity_key[entity_key] = pd.DataFrame(
                 dict(
                     (column_name, simulation.calculate_add(column_name, period = period))
                     for column_name in column_names
                     )
                 )
-        # TODO add roles
-        return openfisca_data_frame_by_entity_key
+            if entity.is_person:
+                person_entity = entity
+            else:
+                non_person_entities.append(entity)
+        if index:
+            person_data_frame = openfisca_data_frame_by_entity_key.get(person_entity.key)
+            if person_data_frame is None:
+                person_data_frame = pd.DataFrame()
+            for entity in non_person_entities:
+                person_data_frame[
+                    "{}_{}".format(entity.key, 'id')
+                    ] = simulation.entities[entity.key].members_entity_id
+                person_data_frame[
+                    "{}_{}".format(entity.key, 'role')
+                    ] = simulation.entities[entity.key].members_legacy_role
+                person_data_frame[
+                    "{}_{}".format(entity.key, 'position')
+                    ] = simulation.entities[entity.key].members_position
+
+        for entity_key, expressions in expressions_by_entity_key.iteritems():
+            data_frame = openfisca_data_frame_by_entity_key[entity_key]
+            for expression in expressions:
+                data_frame[expression] = data_frame.eval(expression)
+
+        if filter_by is not None:
+            openfisca_data_frame_by_entity_key[filter_entity_key] = (
+                openfisca_data_frame_by_entity_key[filter_entity_key].loc[
+                    openfisca_data_frame_by_entity_key[filter_entity_key][filter_by]
+                    ].copy()
+                )
+
+        if not merge:
+            return openfisca_data_frame_by_entity_key
+        else:
+            for entity_key, openfisca_data_frame in openfisca_data_frame_by_entity_key.iteritems():
+                if entity_key != person_entity.key:
+                    openfisca_data_frame.index.name = '{}_id'.format(entity_key)
+                    if len(openfisca_data_frame.reset_index()) > 0:
+                        person_data_frame = person_data_frame.merge(
+                            openfisca_data_frame.reset_index(),
+                            left_on = '{}_id'.format(entity_key),
+                            right_on = '{}_id'.format(entity_key),
+                            )
+            return person_data_frame
 
     def custom_input_data_frame(self, input_data_frame, **kwargs):
         pass
@@ -251,8 +352,6 @@ class AbstractSurveyScenario(object):
     def fill(self, input_data_frame, simulation, period):
         assert period is not None
         log.info('Initialasing simulation using data_frame for period {}'.format(period))
-
-        log.info("Initialasing {}".format(period))
         if period.unit == 'year':  # 1. year
             self.init_simulation_with_data_frame(
                 input_data_frame = input_data_frame,
@@ -274,12 +373,12 @@ class AbstractSurveyScenario(object):
                 simulation = simulation,
                 )
         else:
-            log.info("Unvalid period {}".format(period))
+            log.info("Invalid period {}".format(period))
             raise
 
     def filter_input_variables(self, input_data_frame = None, simulation = None):
         """
-        Clean the data_frame
+        Filter the input data frame from variables that won't be used or are set to be computed
         """
         assert input_data_frame is not None
         assert simulation is not None
@@ -297,14 +396,22 @@ class AbstractSurveyScenario(object):
             role_variable_by_entity_key[entity.key] for entity in simulation.entities.values()
             if not entity.is_person]
 
-        log.info('Variable used_as_input_variables in filter: \n {}'.format(used_as_input_variables))
+        log.debug('Variable used_as_input_variables in filter: \n {}'.format(used_as_input_variables))
+
+        unknown_columns = []
         for column_name in input_data_frame:
             if column_name in id_variables + role_variables:
                 continue
             if column_name not in column_by_name:
-                log.info('Unknown column "{}" in survey, dropped from input table'.format(column_name))
+                unknown_columns.append(column_name)
                 input_data_frame.drop(column_name, axis = 1, inplace = True)
 
+        if unknown_columns:
+            log.debug('The following unknown columns {}, are dropped from input table'.format(
+                sorted(unknown_columns)))
+
+        used_columns = []
+        dropped_columns = []
         for column_name in input_data_frame:
             if column_name in id_variables + role_variables:
                 continue
@@ -316,16 +423,25 @@ class AbstractSurveyScenario(object):
             # Keeping the calculated variables that are initialized by the input data
             if function is not None:
                 if column_name in used_as_input_variables:
-                    log.info(
-                        'Column "{}" not dropped because present in used_as_input_variables'.format(column_name))
+                    used_columns.append(column_name)
                     continue
 
-                log.info('Column "{}" in survey set to be calculated, dropped from input table'.format(column_name))
+                dropped_columns.append(column_name)
                 input_data_frame.drop(column_name, axis = 1, inplace = True)
                 #
             #
         #
-        log.info('Keeping the following variables in the input_data_frame: \n {}'.format(input_data_frame.columns))
+        if used_columns:
+            log.debug(
+                'These columns are not dropped because present in used_as_input_variables:\n {}'.format(
+                    sorted(used_columns)))
+        if dropped_columns:
+            log.debug(
+                'These columns in survey are set to be calculated, we drop them from the input table:\n {}'.format(
+                    sorted(dropped_columns)))
+
+        log.info('Keeping the following variables in the input_data_frame:\n {}'.format(
+            sorted(list(input_data_frame.columns))))
         return input_data_frame
 
     def inflate(self, inflator_by_variable = None, target_by_variable = None):
@@ -386,8 +502,50 @@ class AbstractSurveyScenario(object):
         #
         return self
 
-    def init_simulation_with_data_frame(self, input_data_frame = None, period = None, simulation = None,
-            verbose = False):
+    def init_from_survey_tables(self, calibration_kwargs = None, data_year = None, inflation_kwargs = None,
+            rebuild_input_data = False, rebuild_kwargs = None, input_survey_kwargs = None):
+
+        if data_year is None:
+            data_year = self.year
+
+        if calibration_kwargs is not None:
+            assert set(calibration_kwargs.keys()).issubset(set(
+                ['target_margins_by_variable', 'parameters', 'total_population']))
+
+        if inflation_kwargs is not None:
+            assert set(inflation_kwargs.keys()).issubset(set(['inflator_by_variable', 'target_by_variable']))
+
+        if rebuild_input_data:
+            if rebuild_kwargs is not None:
+                self.build_input_data(year = data_year, **rebuild_kwargs)
+            else:
+                self.build_input_data(year = data_year)
+
+        if self.input_data_table_by_period is None:
+            assert self.input_data_survey_prefix is not None
+            openfisca_survey_collection = SurveyCollection.load(collection = self.collection)
+            openfisca_survey = openfisca_survey_collection.get_survey("{}_{}".format(
+                self.input_data_survey_prefix, data_year))
+            input_data_frame = openfisca_survey.get_values(table = "input").reset_index(drop = True)
+
+            self.init_from_data_frame(
+                input_data_frame = input_data_frame,
+                )
+        else:
+            pass
+        #
+        input_survey_kwargs = input_survey_kwargs if input_survey_kwargs else dict()
+        self.new_simulation(survey = input_survey_kwargs.get('input_survey'))
+        if self.reference_tax_benefit_system is not None:
+            self.new_simulation(reference = True, survey = input_survey_kwargs.get('reference_input_survey'))
+        #
+        if calibration_kwargs:
+            self.calibrate(**calibration_kwargs)
+
+        if inflation_kwargs:
+            self.inflate(**inflation_kwargs)
+
+    def init_simulation_with_data_frame(self, input_data_frame = None, period = None, simulation = None):
         """
         Initialize the simulation period with current input_data_frame
         """
@@ -402,9 +560,12 @@ class AbstractSurveyScenario(object):
         if variables_mismatch:
             log.info(
                 'The following variables used as input variables are not present in the input data frame: \n {}'.format(
-                    variables_mismatch))
-            log.info('The following variables are used as input variables: \n {}'.format(used_as_input_variables))
-            log.info('The input_data_frame contains the following variables: \n {}'.format(input_data_frame.columns))
+                    sorted(variables_mismatch)))
+        if variables_mismatch:
+            log.debug('The following variables are used as input variables: \n {}'.format(
+                sorted(used_as_input_variables)))
+            log.debug('The input_data_frame contains the following variables: \n {}'.format(
+                sorted(list(input_data_frame.columns))))
 
         id_variables = [
             id_variable_by_entity_key[entity.key] for entity in simulation.entities.values()
@@ -441,18 +602,17 @@ class AbstractSurveyScenario(object):
                 continue
             holder = simulation.get_or_new_holder(column_name)
             entity = holder.entity
-            if verbose and (column_serie.values.dtype != holder.column.dtype):
-                log.info(
+            if column_serie.values.dtype != holder.column.dtype:
+                log.debug(
                     'Converting {} from dtype {} to {}'.format(
                         column_name, column_serie.values.dtype, holder.column.dtype)
                     )
             if np.issubdtype(column_serie.values.dtype, np.float):
                 if column_serie.isnull().any():
-                    if verbose:
-                        log.info('There are {} NaN values for {} non NaN values in variable {}'.format(
-                            column_serie.isnull().sum(), column_serie.notnull().sum(), column_name))
-                        log.info('We convert these NaN values of variable {} to {} its default value'.format(
-                            column_name, holder.column.default))
+                    log.debug('There are {} NaN values for {} non NaN values in variable {}'.format(
+                        column_serie.isnull().sum(), column_serie.notnull().sum(), column_name))
+                    log.debug('We convert these NaN values of variable {} to {} its default value'.format(
+                        column_name, holder.column.default))
                     input_data_frame.loc[column_serie.isnull(), column_name] = holder.column.default
                 assert input_data_frame[column_name].notnull().all(), \
                     'There are {} NaN values for {} non NaN values in variable {}'.format(
@@ -473,7 +633,7 @@ class AbstractSurveyScenario(object):
     # def input_data_frame(self):
     #     return self.input_data_frame_by_entity.get(period = periods.period(self.year))
 
-    def new_simulation(self, debug = False, debug_all = False, reference = False, trace = False):
+    def new_simulation(self, debug = False, debug_all = False, reference = False, trace = False, survey = None):
         assert self.tax_benefit_system is not None
         tax_benefit_system = self.tax_benefit_system
         if self.reference_tax_benefit_system is not None and reference:
@@ -502,11 +662,11 @@ class AbstractSurveyScenario(object):
             input_data_frame = self.input_data_frame.copy()
             self.custom_input_data_frame(input_data_frame, period = period)
             self.fill(input_data_frame, simulation, period)
-        # Case 2: fill simulation with a unique input_data_frame containing all entity variables
+        # Case 2: fill simulation with input_data_frame by period containing all entity variables
         elif self.input_data_table_by_period is not None:
             for period, table in self.input_data_table_by_period.iteritems():
                 period = periods.period(period)
-                input_data_frame = self.load_table(table = table)
+                input_data_frame = self.load_table(survey = survey, table = table)
                 self.custom_input_data_frame(input_data_frame, period = period)
                 self.fill(input_data_frame, simulation, period)
         #
@@ -526,7 +686,8 @@ class AbstractSurveyScenario(object):
         survey_collection = SurveyCollection.load(collection = self.collection)
         survey = survey or "{}_{}".format(self.input_data_survey_prefix, self.year)
         survey_ = survey_collection.get_survey(survey)
-        return survey_.get_values(table = table, variables = variables)  # .reset_index(drop = True)
+        log.info("Loading table {} in survey {} from collection {}".format(table, survey, collection))
+        return survey_.get_values(table = table, variables = variables)
 
     def memory_usage(self, reference = False):
         if reference:
@@ -645,6 +806,10 @@ class AbstractSurveyScenario(object):
             elif holder._array_by_period is not None:
                 for period in sorted(holder._array_by_period.keys()):
                     array = holder._array_by_period[period]
+                    if array.shape == ():
+                        print("{}: always = {}".format(period, array))
+                        continue
+
                     print("{}: mean = {}, min = {}, max = {}, mass = {:.2e}, default = {:.1%}, median = {}".format(
                         period,
                         array.mean() if not weighted else np.average(array, weights = weights),
@@ -686,7 +851,7 @@ def init_simulation_with_data_frame_by_entity(input_data_frame_by_entity = None,
             holder = simulation.get_or_new_holder(column_name)
             entity = holder.entity
             if column_serie.values.dtype != holder.column.dtype:
-                log.info(
+                log.debug(
                     'Converting {} from dtype {} to {}'.format(
                         column_name, column_serie.values.dtype, holder.column.dtype)
                     )
@@ -700,3 +865,33 @@ def init_simulation_with_data_frame_by_entity(input_data_frame_by_entity = None,
                 array.size,
                 entity.count)
             holder.array = np.array(array, dtype = holder.column.dtype)
+
+
+# Helpers
+
+def get_words(text):
+    return re.compile('[A-Za-z_]+').findall(text)
+
+
+def assert_variables_in_same_entity(survey_scenario, variables):
+    entity = None
+    for variable_name in variables:
+        variable = survey_scenario.tax_benefit_system.column_by_name.get(variable_name)
+        assert variable
+        if entity is None:
+            entity = variable.entity
+        assert variable.entity == entity, "{} are not from the same entity: {} doesn't belong to {}".format(
+            variables, variable_name, entity.key)
+    return entity.key
+
+
+def get_entity(survey_scenario, variable):
+    variable_ = survey_scenario.tax_benefit_system.column_by_name.get(variable)
+    assert variable_, 'Variable {} is not part of the tax-benefit-system'.format(variable)
+    return variable_.entity
+
+
+def get_weights(survey_scenario, variable):
+    entity = get_entity(survey_scenario, variable)
+    weight_variable = survey_scenario.weight_column_name_by_entity.get(entity.key)
+    return weight_variable